Create Options.xml
Added a series of options to allow the player more control over the difficulty of a given run.
Checkbox each to make gigantic the player's starting equipment (yes) and grenades (no).
Combo each for which who should be able to tinker the gigantic mod (gigantic players) and who should have a lower rarity of the gigantic mod appearing in tables  (gigantic players).